pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpanel-mcp"
version = "0.1.0"
description = "An MCP server to interact with cPanel via API functions to manage email accounts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.11.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
]
keywords = ["cpanel", "email account management", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
cpanel-mcp = "server:run_server"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["config", "cpanel", "routes", "server"]

[project.urls]
Homepage = "https://github.com/your-username/cpanel-mcp"
Repository = "https://github.com/your-username/cpanel-mcp"
Issues = "https://github.com/your-username/cpanel-mcp/issues"
